twitter/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.4.2
Django==2.2.4
django-allauth==0.39.1
django-filter==2.2.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
djoser==2.0.3
drf-nested-routers==0.91
drf-schema-adapter==2.0.6
Flask==1.1.1
idna==2.8
Inflector==3.0.0
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql==0.0.2
mysqlclient==1.4.4
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==6.1.0
psycopg2==2.7.5
psycopg2-binary==2.8.3
PyJWT==1.7.1
python-memcached==1.59
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
redis==3.3.8
requests==2.22.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.15.5
